requirements.txt
Flask==3.0.0
flask-cors==4.0.0
supabase>=2.28.0
PyJWT>=2.10.1
httpx>=0.27.1
python-dotenv==1.0.0
